steering

+++ w5/dungeonGen.cpp
@@ -42,7 +42,7 @@ void gen_drunk_dungeon(char *tiles, size_t w, size_t h)
     // select random point on map
     size_t x = rndWd();
     size_t y = rndHt();
-    startPos.push_back({int(x), int(y)});
+    startPos.push_back({float(x), float(y)});
     size_t numExcavations = 0;
     while (numExcavations < maxExcavations)
     {
